extracted reusable chain config components
