+++ packages/apps/plugins/plugin-chain/src/components/PromptTemplate/PromptTemplate.tsx
@@ -6,7 +6,7 @@ import React, { type PropsWithChildren, useEffect } from 'react';
 
 import { Chain as ChainType } from '@braneframe/types';
 import { getTextContent } from '@dxos/react-client/echo';
-import { DensityProvider, Input, Select, useThemeContext, useTranslation } from '@dxos/react-ui';
+import { DensityProvider, Input, useThemeContext, useTranslation } from '@dxos/react-ui';
 import {
   createBasicExtensions,
   createDataExtensions,
@@ -18,43 +18,8 @@ import { attentionSurface, groupBorder, mx } from '@dxos/react-ui-theme';
 
 import { nameRegex, promptExtension } from './prompt-extension';
 import { CHAIN_PLUGIN } from '../../meta';
-
-const inputTypes = [
-  {
-    value: ChainType.Input.Type.VALUE,
-    label: 'Value',
-  },
-  {
-    value: ChainType.Input.Type.PASS_THROUGH,
-    label: 'Pass through',
-  },
-  {
-    value: ChainType.Input.Type.RETRIEVER,
-    label: 'Retriever',
-  },
-  // {
-  //   value: ChainType.Input.Type.FUNCTION,
-  //   label: 'Function',
-  // },
-  // {
-  //   value: ChainType.Input.Type.QUERY,
-  //   label: 'Query',
-  // },
-  {
-    value: ChainType.Input.Type.RESOLVER,
-    label: 'Resolver',
-  },
-  {
-    value: ChainType.Input.Type.CONTEXT,
-    label: 'Context',
-  },
-  {
-    value: ChainType.Input.Type.SCHEMA,
-    label: 'Schema',
-  },
-];
-
-const getInputType = (type: string) => inputTypes.find(({ value }) => String(value) === type)?.value;
+import { InputRow } from '../InputRow';
+import { TriggerRow } from '../TriggerRow';
 
 const usePromptInputs = (prompt: ChainType.Prompt) => {
   const text = getTextContent(prompt.source) ?? '';
@@ -181,89 +146,6 @@ export const PromptTemplate = ({ prompt }: PromptTemplateProps) => {
         )}
       </div>
     </DensityProvider>
-  );
-};
-
-const TriggerRow = ({ trigger }: { trigger: ChainType.EchoTrigger }) => {
-  return (
-    <tr key={trigger.typename}>
-      <td className='px-3 py-1.5 w-[200px] font-mono text-sm'>{trigger.typename}</td>
-      <td className='px-3 py-1.5 w-[160px]'>{trigger.compareBy}</td>
-      <td className='px-3 py-1.5'>
-        <Input.Root>
-          <Input.TextInput
-            inputMode={'numeric'}
-            classNames={mx('is-full bg-transparent m-2')}
-            value={String(trigger.debounceMs) ?? ''}
-            onChange={(event) => {
-              const parsed = parseInt(event.target.value, 10);
-              trigger.debounceMs = Number.isNaN(parsed) ? 0 : parsed;
-            }}
-          />
-        </Input.Root>
-      </td>
-      <td className='px-3'>
-        <Input.Root>
-          <Input.Checkbox
-            checked={trigger.enabled}
-            onCheckedChange={(checked) => {
-              trigger.enabled = Boolean(checked);
-            }}
-          />
-        </Input.Root>
-      </td>
-    </tr>
-  );
-};
-
-const InputRow = ({ input }: { input: ChainType.Input }) => {
-  const { t } = useTranslation(CHAIN_PLUGIN);
-  return (
-    <tr key={input.name}>
-      <td className='px-3 py-1.5 w-[200px] font-mono text-sm'>{input.name}</td>
-      <td className='px-3 py-1.5 w-[160px]'>
-        <Input.Root>
-          <Select.Root
-            value={String(input.type)}
-            onValueChange={(type) => {
-              input.type = getInputType(type) ?? ChainType.Input.Type.VALUE;
-            }}
-          >
-            <Select.TriggerButton placeholder='Type' classNames='is-full' />
-            <Select.Portal>
-              <Select.Content>
-                <Select.Viewport>
-                  {inputTypes.map(({ value, label }) => (
-                    <Select.Option key={value} value={String(value)}>
-                      {label}
-                    </Select.Option>
-                  ))}
-                </Select.Viewport>
-              </Select.Content>
-            </Select.Portal>
-          </Select.Root>
-        </Input.Root>
-      </td>
-      <td className='px-3'>
-        {[
-          ChainType.Input.Type.VALUE,
-          ChainType.Input.Type.CONTEXT,
-          ChainType.Input.Type.RESOLVER,
-          ChainType.Input.Type.SCHEMA,
-        ].includes(input.type) && (
-          <Input.Root>
-            <Input.TextInput
-              placeholder={t('command placeholder')}
-              classNames={mx('is-full bg-transparent m-2')}
-              value={input.value ?? ''}
-              onChange={(event) => {
-                input.value = event.target.value;
-              }}
-            />
-          </Input.Root>
-        )}
-      </td>
-    </tr>
   );
 };
